Rename route files for clarity
