+++ server/routes/api/v1/apiRouter.js
@@ -1,10 +1,10 @@
 const router = require('express').Router();
 
 router
-    .use(require('./postUsers'))
-    .use(require('./getUsers'))
+    .use(require('./createUser'))
+    .use(require('./getUser'))
     .use(require('./getRoomConfig'))
-    .use(require('./putRoomConfig'))
+    .use(require('./updateRoomConfig'))
     .use(require('./ping'));
 
 module.exports = router;
