+++ server/routes/api/v1/updateRoomConfig.js
